requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
flake8==3.9.2
flake8-isort==4.0.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
inflection==0.5.1
isort==5.9.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==5.4.1
pytest-django==3.9.0
pytz==2020.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
six==1.14.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
testfixtures==6.17.1
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.1.9
zipp==3.1.0
